requirements.txt
asgiref==3.2.3
defusedxml==0.6.0
diff-match-patch==20181111
dj-static==0.0.6
Django==3.0
django-ckeditor==5.8.0
django-import-export==2.0
django-js-asset==1.2.2
djangocms-admin-style==1.4.0
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.2
Pillow==6.2.1
psycopg2==2.8.4
pytz==2019.3
PyYAML==5.2
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0